@@ -1,62 +1,71 @@
 package com.daidaisuki.inventory.controller.dialog;
 
-import com.daidaisuki.inventory.base.controller.BaseDialogController;
-import com.daidaisuki.inventory.model.Customer;
-import com.daidaisuki.inventory.viewmodel.CustomerDialogViewModel;
-import javafx.fxml.FXML;
-import javafx.scene.control.Label;
-import javafx.scene.control.TextField;
 
-public class CustomerDialogController extends BaseDialogController<Customer> {
-  @FXML private TextField nameField;
+/* Comment out during mvvm migration as need to refactor view by view
+
+public class CustomerDialogController
+    extends BaseDialogController<Customer, CustomerDialogViewModel> {
+
+  @FXML private TextField fullNameField;
   @FXML private TextField phoneNumberField;
   @FXML private TextField emailField;
   @FXML private TextField addressField;
-  @FXML private TextField platformField;
+  @FXML private TextField acquisitionSourceField;
+
   @FXML private Label dialogTitle;
 
-  private CustomerDialogViewModel viewModel;
+  public CustomerDialogController(CustomerDialogViewModel viewModel) {
+    super(viewModel);
+  }
+
+  // Comment out during mvvm migration as need to refactor view by view
+
+  @Override
+  public Customer getModel() {
+    return this.viewModel.getModel();
+  }
 
   @Override
   public void setModel(Customer customer) {
-    this.viewModel = new CustomerDialogViewModel(customer);
-    this.model = customer;
-    boolean isEdit = customer != null;
-    String title = isEdit ? "Edit Customer" : "Add Customer";
-    dialogStage.setTitle(title);
-    dialogTitle.setText(title);
-    if (isEdit) {
-      nameField.setText(customer.getName());
-      phoneNumberField.setText(customer.getPhoneNumber());
-      emailField.setText(customer.getEmail());
-      addressField.setText(customer.getAddress());
-      platformField.setText(customer.getPlatform());
-    }
+    this.viewModel.setModel(customer);
+    setupBinding();
+    setupBaseBinding();
+    this.dialogTitle
+        .textProperty()
+        .bind(
+            Bindings.when(this.viewModel.isNewProperty())
+                .then("Create Customer")
+                .otherwise("Edit Customer"));
+    this.dialogStage
+        .titleProperty()
+        .bind(
+            Bindings.when(this.viewModel.isNewProperty())
+                .then("Add Customer")
+                .otherwise("Edit Customer"));
+  }
+
+  private void setupBinding() {
+    this.phoneNumberField.setTextFormatter(
+        new TextFormatter<>(
+            change -> change.getControlNewText().matches("\\+?[0-9]{0-15}") ? change : null));
+    this.fullNameField.textProperty().bindBidirectional(this.viewModel.fullName);
+    this.phoneNumberField.textProperty().bindBidirectional(this.viewModel.phoneNumber);
+    this.emailField.textProperty().bindBidirectional(this.viewModel.email);
+    this.addressField.textProperty().bindBidirectional(this.viewModel.address);
+    this.acquisitionSourceField.textProperty().bindBidirectional(this.viewModel.acquisitionSource);
   }
 
   @FXML
   @Override
   protected void handleSave() {
-    String name = sanitizeInput(nameField);
-    String phoneNumber = sanitizeOrNull(phoneNumberField);
-    String email = sanitizeOrNull(emailField);
-    String address = sanitizeOrNull(addressField);
-    String platform = sanitizeInput(platformField);
-
-    String errorMessage =
-        viewModel.validateAndUpdateCustomer(name, phoneNumber, email, address, platform);
-
-    if (errorMessage.isEmpty()) {
-      showError(errorMessage);
-      return;
+    try {
+      this.viewModel.save();
+      this.saveClicked = true;
+      closeDialog();
+    } catch (Exception e) {
+      showError("Failed to save customer: \n" + e.getMessage());
     }
-
-    if (model == null) {
-      model = new Customer();
-    }
-
-    model = viewModel.getCustomer();
-    saveClicked = true;
-    dialogStage.close();
   }
+  //
 }
+*/
