@@ -2,93 +2,75 @@ package com.daidaisuki.inventory.controller.dialog;
 
 import com.daidaisuki.inventory.base.controller.BaseDialogController;
 import com.daidaisuki.inventory.model.Product;
-import com.daidaisuki.inventory.service.ProductService;
-import com.daidaisuki.inventory.util.ValidationUtils;
-import java.sql.SQLException;
+import com.daidaisuki.inventory.viewmodel.dialog.ProductDialogViewModel;
+import java.util.List;
+import javafx.beans.binding.Bindings;
+import javafx.collections.FXCollections;
 import javafx.fxml.FXML;
+import javafx.scene.control.CheckBox;
+import javafx.scene.control.ComboBox;
 import javafx.scene.control.Label;
 import javafx.scene.control.TextField;
 
-public class ProductDialogController extends BaseDialogController<Product> {
-
+public class ProductDialogController extends BaseDialogController<Product, ProductDialogViewModel> {
+  @FXML private TextField skuField;
   @FXML private TextField nameField;
   @FXML private TextField categoryField;
-  @FXML private TextField stockField;
+  @FXML private TextField descriptionField;
+  @FXML private TextField weightField;
   @FXML private TextField priceField;
-  @FXML private TextField costField;
-  @FXML private TextField shippingField;
+
+  @FXML private ComboBox<String> unitTypeComboBox;
+
+  @FXML private CheckBox activeCheckBox;
 
   @FXML private Label dialogTitle;
 
-  private ProductService productService;
+  public ProductDialogController(ProductDialogViewModel dialogViewModel) {
+    super(dialogViewModel);
+  }
 
-  public void setProductService(ProductService productService) {
-    this.productService = productService;
+  @FXML
+  public void initialize() {
+    // Setting items first then followed by binding
+    List<String> existingUnitTypes = this.viewModel.getAvailableUnitTypes();
+    this.unitTypeComboBox.setEditable(true);
+    this.unitTypeComboBox.setItems(FXCollections.observableArrayList(existingUnitTypes));
+    setupBinding();
+    this.dialogTitle
+        .textProperty()
+        .bind(
+            Bindings.when(this.viewModel.isNewProperty())
+                .then("Create Product")
+                .otherwise("Edit Product"));
+    if (confirmButton != null) {
+      confirmButton.disableProperty().bind(this.viewModel.isInvalidProperty());
+    }
   }
 
   @Override
-  public void setModel(Product product) {
-    this.model = product;
-    boolean isEdit = product != null;
-    String title = isEdit ? "Edit Product" : "Add Product";
-    dialogStage.setTitle(title);
-    dialogTitle.setText(title);
-    if (isEdit) {
-      nameField.setText(product.getName());
-      categoryField.setText(product.getCategory());
-      stockField.setText(String.valueOf(product.getStock()));
-      priceField.setText(String.valueOf(product.getPrice()));
-      costField.setText(String.valueOf(product.getCost()));
-      shippingField.setText(String.valueOf(product.getShipping()));
-    }
+  public Product getResult() {
+    return confirmed ? this.viewModel.createResult() : null;
+  }
+
+  private void setupBinding() {
+    this.skuField.textProperty().bindBidirectional(this.viewModel.sku);
+    this.nameField.textProperty().bindBidirectional(this.viewModel.name);
+    this.categoryField.textProperty().bindBidirectional(this.viewModel.category);
+    this.descriptionField.textProperty().bindBidirectional(this.viewModel.description);
+    this.weightField.textProperty().bindBidirectional(this.viewModel.weight);
+    this.priceField.textProperty().bindBidirectional(this.viewModel.price);
+    this.unitTypeComboBox.valueProperty().bindBidirectional(this.viewModel.unitType);
+    this.activeCheckBox.selectedProperty().bindBidirectional(this.viewModel.isActive);
   }
 
   @FXML
   @Override
-  protected void handleSave() {
-    StringBuilder errorMessage = new StringBuilder();
-    String name = sanitizeInput(nameField);
-    String category = sanitizeInput(categoryField);
-    String stockStr = sanitizeInput(stockField);
-    String priceStr = sanitizeInput(priceField);
-    String costStr = sanitizeInput(costField);
-    String shippingStr = sanitizeInput(shippingField);
-
-    ValidationUtils.isFieldEmpty(name, "Name", errorMessage);
-    ValidationUtils.isFieldEmpty(category, "Category", errorMessage);
-    ValidationUtils.isNumeric(stockStr, "Stock Number", errorMessage, false);
-    ValidationUtils.isNumeric(priceStr, "Sale Price", errorMessage, true);
-    ValidationUtils.isNumeric(costStr, "Purchase Cost", errorMessage, true);
-    ValidationUtils.isNumeric(shippingStr, "Shipping Fee", errorMessage, true);
-
-    if (errorMessage.length() > 0) {
-      showError(errorMessage.toString());
-      return;
+  protected void handleConfirm() {
+    if (!this.viewModel.isInvalidProperty().get()) {
+      this.confirmed = true;
+      this.dialogStage.close();
+    } else {
     }
-
-    if (model == null) {
-      model = new Product();
-    }
-
-    model.setName(name);
-    model.setCategory(category);
-    model.setStock(Integer.parseInt(stockStr));
-    model.setPrice(Double.parseDouble(priceStr));
-    model.setCost(Double.parseDouble(costStr));
-    model.setShipping(Double.parseDouble(shippingStr));
-
-    try {
-      if (model.getId() == 0) {
-        productService.addProduct(model);
-      } else {
-        productService.updateProduct(model);
-      }
-    } catch (SQLException e) {
-      showError("Failed to save product: \n" + e.getMessage());
-      return;
-    }
-
-    saveClicked = true;
-    dialogStage.close();
   }
 }
