@@ -1,7 +1,11 @@
 package com.daidaisuki.inventory.exception;
 
-public class InsufficientStockException extends Exception {
+public class InsufficientStockException extends RuntimeException {
   public InsufficientStockException(String message) {
     super(message);
+  }
+
+  public InsufficientStockException(String message, Throwable cause) {
+    super(message, cause);
   }
 }
